+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/ui/favorite/FavoriteVIewModel.kt
@@ -2,15 +2,12 @@ package com.pinheiro.marvelhqs.presenter.ui.favorite
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewModelScope
 import com.pinheiro.marvelhqs.domain.mapper.comicDTOListTOComicViewObjectList
 import com.pinheiro.marvelhqs.domain.usecase.GetFavoriteUseCase
 import com.pinheiro.marvelhqs.domain.viewobject.ComicViewObject
-import com.pinheiro.marvelhqs.presenter.DefaultPaginator
 import com.pinheiro.marvelhqs.presenter.ScreenState
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
