+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/ui/navigation/ComicNavigation.kt
@@ -1,7 +1,6 @@
 package com.pinheiro.marvelhqs.presenter.ui.navigation
 
 import android.annotation.SuppressLint
-import android.content.Context
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
@@ -10,16 +9,11 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.FavoriteBorder
 import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.icons.filled.Output
 import androidx.compose.material.icons.outlined.Home
-import androidx.compose.material.icons.outlined.Info
-import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalDrawerSheet
 import androidx.compose.material3.ModalNavigationDrawer
@@ -28,7 +22,6 @@ import androidx.compose.material3.NavigationDrawerItemDefaults
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -37,22 +30,22 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
+import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.pinheiro.marvelhqs.R
 import com.pinheiro.marvelhqs.presenter.ui.comic.ComicScreen
+import com.pinheiro.marvelhqs.presenter.ui.components.MarvelTopAppBar
 import com.pinheiro.marvelhqs.presenter.ui.favorite.FavoriteScreen
 import kotlinx.coroutines.launch
-import kotlinx.serialization.Serializable
 
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ComicNavigation() {
+fun ComicNavigation(mainNavController: NavHostController) {
     val items = listOf(
         NavigationItem(
             title = "All",
@@ -62,17 +55,17 @@ fun ComicNavigation() {
 
         ),
         NavigationItem(
-            title = "Favorite",
+            title = stringResource(R.string.FAVORITE),
             selectedIcon = Icons.Default.Favorite,
             unselectedIcon = Icons.Default.FavoriteBorder,
-            badgeCount = 45,
             route = "favorite"
         ),
-//        NavigationItem(
-//            title = "Settings",
-//            selectedIcon = Icons.Filled.Settings,
-//            unselectedIcon = Icons.Outlined.Settings,
-//        ),
+        NavigationItem(
+            title = stringResource(R.string.LOG_OUT),
+            selectedIcon = Icons.Default.Output,
+            unselectedIcon = Icons.Default.Output,
+            route = "exit"
+        ),
     )
 
     val navController = rememberNavController()
@@ -130,24 +123,9 @@ fun ComicNavigation() {
         ) {
             Scaffold(
                 topBar = {
-                    TopAppBar(
-                        title = {
-                            Text(text = "Marvel")
-                        },
-                        navigationIcon = {
-                            IconButton(onClick = {
-                                scope.launch {
-                                    drawerState.open()
-                                }
-                            }) {
-                                Icon(
-                                    imageVector = Icons.Default.Menu,
-                                    contentDescription = "Menu"
-                                )
-                            }
-                        }
-                    )
-                }
+                    MarvelTopAppBar(drawerState)
+        }
+
             ) {
 
                 NavHost(navController = navController, startDestination = "home") {
@@ -157,6 +135,13 @@ fun ComicNavigation() {
                     composable("favorite") {
                         FavoriteScreen()
                     }
+                    composable("exit") {
+                        navController.popBackStack()
+                        mainNavController.navigate(Login) {
+                            popUpTo(0)
+                        }
+
+                    }
                 }
 
             }
@@ -164,24 +149,12 @@ fun ComicNavigation() {
     }
 }
 
-@Preview
-@Composable
-fun NavigatorPreview() {
-//    ComicNavigation()
-}
 
-
-data class NavigationItem(
-    val title: String,
-    val selectedIcon: ImageVector,
-    val unselectedIcon: ImageVector,
-    val badgeCount: Int? = null,
-    val route: String,
-
-    )
-
-@Serializable
-object Favorite
-
-@Serializable
-object All
+//data class NavigationItem(
+//    val title: String,
+//    val selectedIcon: ImageVector,
+//    val unselectedIcon: ImageVector,
+//    val badgeCount: Int? = null,
+//    val route: String,
+//
+//    )
