+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/ui/navigation/MainNavigationRouts.kt
@@ -1,0 +1,15 @@
+package com.pinheiro.marvelhqs.presenter.ui.navigation
+
+import kotlinx.serialization.Serializable
+@Serializable
+object Logged
+
+@Serializable
+object Home
+@Serializable
+object Login
+@Serializable
+object Registration
+
+@Serializable
+object Comics
